:recycle: Add comment component.
